feat: Scanner API initialized
