+++ core/src/main/java/org/kybinfrastructure/Main.java
@@ -1,6 +1,6 @@
 package org.kybinfrastructure;
 
-import org.kybinfrastructure.ioc.KybContainer;
+import org.kybinfrastructure.ioc.KybContainerBuilder;
 import org.kybinfrastructure.ioc.test_classes.SomeImpl;
 import java.util.ArrayList;
 import java.util.List;
@@ -10,7 +10,7 @@ public final class Main {
 
 	public static void main(String[] args) throws ClassNotFoundException {
 		System.out.println("Hello World!");
-		var container = KybContainer.build(Main.class);
+		var container = KybContainerBuilder.build(Main.class);
 		SomeImpl initialized = container.getImpl(SomeImpl.class);
 		initialized.sayHi("Onur");
 	}
